doing some changes, refactoring, testiting again

+++ src/main/java/com/filmstar/api/controllers/commands/DoUpdateMoviePutController.java
@@ -41,7 +41,7 @@ public class DoUpdateMoviePutController {
      * @return ResponseEntity con información sobre el resultado de la operación de actualización.
      */
     @PutMapping
-    public ResponseEntity<?> execute(@PathVariable Integer id, 
+    public ResponseEntity<?> execute(@PathVariable Long id, 
                                      @Valid @RequestBody Movie updatedMovie,
                                      BindingResult bindingResult) {
         ResponseEntity<?> response = updateMovieUseCase.execute(id, updatedMovie, bindingResult);
